Extract category stat chart into its own file
